cambios

+++ routes/web.php
@@ -16,31 +16,33 @@ Route::get('/', function () {
 });
 
 Auth::routes();
+//rutas de prueba
 Route::get('product', 'TestController@index');
+Route::get('product/{id}', 'ProductController@show');
+//Ruta de carrito de compras
+Route::post('/cart', 'CartDetailController@store');
 
 
 Route::get('/home', 'HomeController@index')->name('home');
 
 Route::middleware(['auth', 'admin'])->prefix('')->group(function () {
     //rutas de productos
-    Route::get('Admin/product', 'ProductController@index');
-    Route::get('Admin/product/create', 'ProductController@create');
-    Route::get('Admin/product/create', 'ProductController@traercat');
-    Route::post('Admin/product', 'ProductController@store');
-    Route::get('Admin/product/{id}/edit', 'ProductController@edit');
-    Route::post('Admin/product/{id}/edit', 'ProductController@update');
-    Route::put('Admin/activar/{id}', 'ProductController@activar');
-    Route::put('Admin/desactivar/{id}', 'ProductController@desactivar');
+    Route::get('Admin/product', 'Admin\ProductController@index');
+    Route::get('Admin/product/create', 'Admin\ProductController@create');
+    Route::get('Admin/product/create', 'Admin\ProductController@traercat');
+    Route::post('Admin/product', 'Admin\ProductController@store');
+    Route::get('Admin/product/{id}/edit', 'Admin\ProductController@edit');
+    Route::post('Admin/product/{id}/edit', 'Admin\ProductController@update');
+    Route::put('Admin/activar/{id}', 'Admin\ProductController@activar');
+    Route::put('Admin/desactivar/{id}', 'Admin\ProductController@desactivar');
     //images con productos
-    Route::get('Admin/product/{id}/images', 'ImageController@index');//listar
-    Route::post('Admin/product/{id}/images', 'ImageController@store');//crear
-    Route::delete('Admin/product/{id}/images', 'ImageController@destroy');//eliminar  
-    Route::get('Admin/product/{id}/images/select/{image}', 'ImageController@select');//destacar   
-
+    Route::get('Admin/product/{id}/images', 'Admin\ImageController@index');//listar
+    Route::post('Admin/product/{id}/images', 'Admin\ImageController@store');//crear
+    Route::delete('Admin/product/{id}/images', 'Admin\ImageController@destroy');//eliminar  
+    Route::get('Admin/product/{id}/images/select/{image}', 'Admin\ImageController@select');//destacar   
 
 
     //rutas de categorias
     Route::get('Admin/category', 'CategoryController@index');
 });
 
-
